Rename AdminPanel.tsx to AdminPage.tsx
